requirements.txt
aiohttp~=3.7.2
aiohttp_cors~=0.7.0
asyncpg~=0.21.0
git+https://github.com/ckkz-it/databases.git@sqlalchemy-defaults#egg=databases[postgresql]
marshmallow~=3.9.1
SQLAlchemy~=1.3.20

pytest~=6.1.2
pytest-aiohttp~=0.3.0
pytest-cov~=2.10.1
python-dotenv~=0.15.0
flake8~=3.8.4
isort~=5.6.4
flake8-quotes~=3.2.0
